backend/requirements.txt
flask
spotipy
python-dotenv
pyjwt